+++ src/test/groovy/groovyx/net/http/HttpResponseException.groovy
@@ -1,0 +1,12 @@
+package groovyx.net.http
+
+class HttpResponseException extends RuntimeException {
+    final int statusCode
+    final Object response
+
+    HttpResponseException(int statusCode, Object response = null, String message = null) {
+        super(message ?: "HTTP ${statusCode}")
+        this.statusCode = statusCode
+        this.response = response
+    }
+}